refactor(hub-common): add "/helpers/format-fetchers" subfolder to improve organization
affects: @esri/hub-common
