+++ packages/common/src/downloads/helpers/format-fetchers/getAllExportItemFormats.ts
@@ -1,4 +1,4 @@
-import { IDynamicDownloadFormat, PORTAL_EXPORT_TYPES } from "./types";
+import { IDynamicDownloadFormat, PORTAL_EXPORT_TYPES } from "../types";
 
 export function getAllExportItemFormats(): IDynamicDownloadFormat[] {
   return Object.keys(PORTAL_EXPORT_TYPES).map((format) => ({
