+++ packages/common/src/downloads/helpers/format-fetchers/getHubDownloadApiFormats.ts
@@ -1,9 +1,9 @@
-import { IArcGISContext } from "../../ArcGISContext";
-import { isHostedFeatureServiceEntity } from "../../content";
-import { IHubEditableContent } from "../../core/types/IHubEditableContent";
+import { IArcGISContext } from "../../../ArcGISContext";
+import { isHostedFeatureServiceEntity } from "../../../content";
+import { IHubEditableContent } from "../../../core/types/IHubEditableContent";
 import { getCreateReplicaFormats } from "./getCreateReplicaFormats";
 import { getPagingJobFormats } from "./getPagingJobFormats";
-import { IDynamicDownloadFormat } from "./types";
+import { IDynamicDownloadFormat } from "../types";
 
 export function getHubDownloadApiFormats(
   entity: IHubEditableContent
